Add API response for single organisation
This one shows all the documents associated with the organisation in a 'contents' array. There's probably a bit more work to do on naming things here. I also still need to wire in some other relationships and make the code generate the proper API URLs. I'm leaving the latter until the API can handle a few more types of content as that'll be the point at which the URL structure becomes clearer.